pyproject.toml
[tool.poetry]
name = "zoo-management-api"
version = "0.1.0"
description = "Zoo Management API with FastAPI and GraphQL"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
strawberry-graphql = "^0.211.1"
motor = "^3.3.1"
redis = "^5.0.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"
cryptography = "^41.0.5"
beanie = "^1.23.0"
gunicorn = "^21.2.0"
elasticsearch = "^7.17.0"
elasticsearch-dsl = "^7.4.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"